Ported Lyrics, WatchVideo, Selector to uhtml
