+++ package.json
@@ -15,7 +15,7 @@
   "devDependencies": {
     "@netlify/blobs": "^9.1.1",
     "@netlify/edge-functions": "^2.12.0",
-    "@types/node": "^22.15.17",
+    "@types/node": "^22.15.18",
     "@types/sortablejs": "^1.15.8",
     "autoprefixer": "^10.4.21",
     "eruda": "^3.4.1",
