+++ src/components/UpdatePrompt.ts
@@ -1,12 +1,11 @@
 import './UpdatePrompt.css';
-import { i18n } from "../lib/utils";
-import { html } from 'uhtml';
+import { i18n } from '../lib/utils';
+import { html, render } from 'uhtml';
 
-export default async function UpdatePrompt() {
+export default async function UpdatePrompt(dialog: HTMLDialogElement) {
 
   const commitsSrc = 'https://api.github.com/repos/n-ce/ytify/commits/main';
   const commitsLink = 'https://github.com/n-ce/ytify/commits';
-
   const list = await fetch(commitsSrc)
     .then(res => res.json())
     .then(data => data.commit.message.split('-'))
@@ -14,19 +13,22 @@ export default async function UpdatePrompt() {
     .catch(() => html`<li>Failed to load update data from Github.</li>`);
 
 
-  return html`
+  render(dialog, html`
     <ul>
       ${list}
       <hr />
-      <li @click=${() => { open(commitsLink) }}>
+      <li @click=${() => {
+      open(commitsLink);
+    }}>
         ${i18n('updater_changelog_full')}
       </li>
     </ul>
     <span>
       <button id="updateBtn" autofocus>
-      ${i18n('updater_update')}
+        ${i18n('updater_update')}
       </button>
       <button id="laterBtn">${i18n('updater_later')}</button>
-    </span>`;
+    </span>
+    `);
 
 }
