+++ src/main.ts
@@ -6,7 +6,6 @@ import './scripts/list';
 import './scripts/search';
 import './scripts/library';
 import { render } from 'solid-js/web';
-import { render as uhtml } from 'uhtml';
 import { actionsMenu, superCollectionList } from './lib/dom';
 
 addEventListener('DOMContentLoaded', async () => {
@@ -43,8 +42,7 @@ addEventListener('DOMContentLoaded', async () => {
           })
 
           import('./components/UpdatePrompt')
-            .then(async mod => uhtml(dialog, await mod.default()))
-            .then(() => document.body.appendChild(dialog));
+            .then(mod => mod.default(dialog))
         }
       });
     });
