+++ src/modules/start.ts
@@ -4,7 +4,6 @@ import { $, getDownloadLink, i18n, idFromURL, proxyHandler } from '../lib/utils'
 import { bitrateSelector, searchFilters, superInput, audio, loadingScreen, ytifyIcon, searchlist } from '../lib/dom';
 import fetchList from '../modules/fetchList';
 import { fetchCollection } from "../lib/libraryUtils";
-import { render } from 'solid-js/web';
 
 export default async function() {
 
@@ -61,8 +60,12 @@ export default async function() {
     loadingScreen.showModal();
     if (isPWA && shareAction === 'watch') {
       store.actionsMenu.id = id;
-      import('../components/WatchVideo')
-        .then(mod => render(mod.default, document.body));
+      const dialog = document.createElement('dialog') as HTMLDialogElement;
+      dialog.open = true;
+      dialog.className = 'watcher';
+      document.body.appendChild(dialog);
+      import('../components/WatchVideo.ts')
+        .then(mod => mod.default(dialog));
     }
     else if (isPWA && shareAction) {
       const a = $('a');
